(feat) minor changes, replace ArrayDeque in BallotBox
