+++ jraft-core/src/main/java/com/alipay/sofa/jraft/util/internal/ReflectionIntegerFieldUpdater.java
@@ -14,38 +14,38 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.jraft.rhea.util.internal;
+package com.alipay.sofa.jraft.util.internal;
 
 import java.lang.reflect.Field;
-
-import sun.misc.Unsafe;
 
 /**
  *
  * @author jiachun.fjc
  */
-@SuppressWarnings("unchecked")
-final class UnsafeReferenceFieldUpdater<U, W> implements ReferenceFieldUpdater<U, W> {
+final class ReflectionIntegerFieldUpdater<U> implements IntegerFieldUpdater<U> {
 
-    private final long   offset;
-    private final Unsafe unsafe;
+    private final Field field;
 
-    UnsafeReferenceFieldUpdater(Unsafe unsafe, Class<? super U> tClass, String fieldName) throws NoSuchFieldException {
-        final Field field = tClass.getDeclaredField(fieldName);
-        if (unsafe == null) {
-            throw new NullPointerException("unsafe");
+    ReflectionIntegerFieldUpdater(Class<? super U> tClass, String fieldName) throws NoSuchFieldException {
+        this.field = tClass.getDeclaredField(fieldName);
+        this.field.setAccessible(true);
+    }
+
+    @Override
+    public void set(final U obj, final int newValue) {
+        try {
+            this.field.set(obj, newValue);
+        } catch (final IllegalAccessException e) {
+            throw new RuntimeException(e);
         }
-        this.unsafe = unsafe;
-        this.offset = unsafe.objectFieldOffset(field);
     }
 
     @Override
-    public void set(final U obj, final W newValue) {
-        this.unsafe.putObject(obj, this.offset, newValue);
-    }
-
-    @Override
-    public W get(final U obj) {
-        return (W) this.unsafe.getObject(obj, this.offset);
+    public int get(final U obj) {
+        try {
+            return (Integer) this.field.get(obj);
+        } catch (final IllegalAccessException e) {
+            throw new RuntimeException(e);
+        }
     }
 }
