+++ jraft-core/src/main/java/com/alipay/sofa/jraft/util/internal/ReflectionReferenceFieldUpdater.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.jraft.rhea.util.internal;
+package com.alipay.sofa.jraft.util.internal;
 
 import java.lang.reflect.Field;
 
@@ -22,17 +22,18 @@ import java.lang.reflect.Field;
  *
  * @author jiachun.fjc
  */
-final class ReflectionLongFieldUpdater<U> implements LongFieldUpdater<U> {
+@SuppressWarnings("unchecked")
+final class ReflectionReferenceFieldUpdater<U, W> implements ReferenceFieldUpdater<U, W> {
 
     private final Field field;
 
-    ReflectionLongFieldUpdater(Class<? super U> tClass, String fieldName) throws NoSuchFieldException {
+    ReflectionReferenceFieldUpdater(Class<? super U> tClass, String fieldName) throws NoSuchFieldException {
         this.field = tClass.getDeclaredField(fieldName);
         this.field.setAccessible(true);
     }
 
     @Override
-    public void set(final U obj, final long newValue) {
+    public void set(final U obj, final W newValue) {
         try {
             this.field.set(obj, newValue);
         } catch (final IllegalAccessException e) {
@@ -41,9 +42,9 @@ final class ReflectionLongFieldUpdater<U> implements LongFieldUpdater<U> {
     }
 
     @Override
-    public long get(final U obj) {
+    public W get(final U obj) {
         try {
-            return (Long) this.field.get(obj);
+            return (W) this.field.get(obj);
         } catch (final IllegalAccessException e) {
             throw new RuntimeException(e);
         }
